requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.23.10
botocore==1.26.10
cycler==0.11.0
Django==4.0.4
django-cors-headers==3.12.0
django-storages==1.12.3
djangorestframework==3.13.1
fonttools==4.33.3
gunicorn==20.1.0
jmespath==1.0.0
kiwisolver==1.4.2
matplotlib==3.5.2
numpy==1.22.4
packaging==21.3
Pillow==9.1.1
PyJWT==2.4.0
pymongo==4.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
